MULE-10937: Separate deployment artifact implementations from deployment module
_ Applying changes from code review
