+++ modules/deployment-impl/src/test/java/org/mule/runtime/module/deployment/impl/internal/artifact/DefaultArtifactClassLoaderManagerTestCase.java
@@ -5,7 +5,7 @@
  * LICENSE.txt file.
  */
 
-package org.mule.runtime.module.deployment.internal;
+package org.mule.runtime.module.deployment.impl.internal.artifact;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
