+++ modules/launcher/pom.xml
@@ -66,7 +66,7 @@
         </dependency>
         <dependency>
             <groupId>org.mule.modules</groupId>
-            <artifactId>mule-module-deployment-impl</artifactId>
+            <artifactId>mule-module-deployment-model-impl</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
             <type>test-jar</type>
