+++ modules/pom.xml
@@ -20,8 +20,8 @@
         <module>cxf</module>
         <module>container</module>
         <module>deployment</module>
-        <module>deployment-impl</module>
         <module>deployment-model</module>
+        <module>deployment-model-impl</module>
         <module>dsl-api</module>
         <module>extensions-support</module>
         <module>file-extension-common</module>
